Fix support for directional lights (#9)
The default light is a directional one, so we must initialize its
position (which, according to the standard, is (0, 0, 1)) following the
same algorithm we apply when glLightfv() is called for directional
lights: we multiply its position by 100000. The reason why we do this is
that GX does not support ideal directional lights, so we simulate them
by placing the light very far away and disabling the distance
attenuation.

Also, we don't need to set the light direction on the GX light, since
the light is positioned far away and that by itself will determine under
which angle our objects will be lit.