Implement support for specular lights
We add the specular lights to the first channel, along with ambient
lights: ambient lights are not directional (they only depend on the
distance), so we can switch the angular attenuation function to
GX_AF_SPEC without affecting them.

Note that there's a difference between how GX and OpenGL compute
specular lights: in OpenGL, it is a combination that depends solely
on the GL_SPECULAR values set on the light and on the material; but in
GX, due to the fact that we have only two channels for mixing lights, we
pre-multiply the two GL_SPECULAR values of the light and of the
material, and then we have an additional factor which is the material
ambient color and which we cannot avoid. The reason is that GX computes
the resulting color of the specular light by mixing the color set on the
light with the object's material.