glDraw{Arrays,Elements}: don't optimize away texture coords when in call lists
When recording a call list we don't know if the list will be played back
with texturing enabled or not, so, if we have got texture coordinates,
place them into the GX list regardless of the current texturing state.

Conversely, when setting up the TEV rendering stages don't worry about
vertex data, but just on whether texturing is enabled. The reason is
that when playing back a display list we do not know what vertex data
is present in it (we might add the information in the list in the form
of flags, but this is an optimisation we can do later).

We keep the optimization only when directly drawing.