Textures: let client choose which fast conversions to activate
This is a compromise to allow keeping the binary size small while still
activating all the faster textures conversions needed: we instantiate all
the texture conversion templates, but we don't directly use them, so
that the linker will not bring them into the compiled object. Instead,
we expose their symbols and a public function to let the client register
the conversions it needs.

The conversion functions are exposed as intptr_t in order to hide the
full signature, since we might need to change it later.

If the client doesn't do this, opengx will pick the slower conversion
(which can be up to 10x slower), but the program will not break in any
case.