textures: allow binding textures with non generated names
Calling glGenTextures() is not necessary in order to use a texture name:
any number greater than 0 should work. So, drop the requirement for
texture names to be reserved by glGenTextures(), and do the reservation
in glBindTexture() itself, if needed.

This also means that we should rework our texture storage, and replace
the array with a map. But let's leave that for the future.