Fix stride computation for glDrawArray()
The stride passed by the client is in bytes, while the current
implementation stores the stride as the number of floats to skip. So, we
need to divide the value passed by the client by sizeof(float).

Note that there a couple of big issues with the current implementation,
that will need to be fixed:
- the `type` parameter is ignored, we always assume that the client
  passes floats.
- the stride value is assumed to be a multiple of 4 (`sizeof(float)`),
  whereas the client could pass any value.