projection: fix glOrtho()'s matrix transposition
The memory layout of the matrices handled by glMultMatrix and Mtx44 is
different. We need to transpose the matrix, if we store it as a Mtx44.

This was causing the view to always being centered on the screen, even
when the horizontal and vertical coordinates passed to gluOrtho2D were
not symmetrical.