Implement glAlphaFunc() and alpha test (#45)
Contrary to what the comment was stating, there is no need for an
additional TEV stage in order to implement alpha testing. It can be done
almost trivially, since the GX API supports it. The only trick is to
move the Z-comparison step after the texturing, and remember resetting
it back during glClear().