geometry: normalize matrix in glMultMatrix()
The matrix passed by the client might have its fourth row set to
something other than the usual [0, 0, 0, 1]: for example, the dinoshade
example program[1] computes a shadow matrix whose last row is [0, 0, 0,
8000]. This creates a problem, because the GX API only expects 3x4
matrices for the model-view transformation, so if we just cut off the
last row we produce a wrong transformation.

While we cannot solve this problem in the general case, in practice the
first three elements of the last row will always be zeroes, so we can
just focus on the last element: if it's different than 1, we can scale
the hole matrix by dividing it by this value and obtain a 4x4 matrix
where the last row is the usual [0, 0, 0, 1], which GX can handle.

[1]: https://github.com/markkilgard/glut/blob/master/progs/examples/dinoshade.c