stencil: update stencil buffer before main drawing
Draw to the stencil buffer before drawing the graphic primitive,
otherwise the stencil drawing operations which perform a Z-buffer test
(the `zpass` and `zfail` ones set with glStencilOp) will be affected by
it.
When moving the stencil drawing operations before the main drawing,
special care must be taken to avoid using the updated stencil buffer
already when drawing the current primitive (which should be drawn using
the old stencil buffer). We use the `draw_count` variable to achieve
this.