Minimal implementation of glPolygonOffset()
Implement support for the `offset` parameter of glPolygonOffset() (the
`factor` parameter requires analyzing the vertices' coordinates and some
more complex computation), to add an offset to the Z coordinates. We do
this by altering the projection matrix. The 0.00001f has been found
empirically, it may be that in the future we need to tweak it or compute
it at runtime.