Implement glReadBuffer() and glDrawBuffer()
We don't really know if double buffering is used, since it's a decision
that the integration library takes.

We limit the implementation to a check whether the correct value is
being set; we don't really allow changing the read or the write buffer,
since reading or writing from the XFB would be troublesome (it's in YUV
format).