Fix 2D projection to show the full (0,0)-(639,479) rectangle
The transformation was incorrect, because it was causing the row 0 and
the column 0 to fall outside of the screen: if we want the leftmost
column of screen pixels to show the scene pixels having coordinate 0,
then the left coordinate to guOrtho() must be -0.5, and in this way a
scene pixel having 0 as the x coordinate will be centered on the
leftmost pixel.

For the rightmost column the computation was correct (the value we pass
to guOrtho() must be 639.5, so we remove the -1 addend in order to
maintain the old value.