TEV: implement full support for texture matrices
With this commit we actually use the texture matrices and pass them to
the GPU. Doing this, we "break" the functionality of the
dirty_texture_gen bit, in the sense that we rebuild the texture
generation matrices regardless of whether the dirty flag is set. We
might bring this optimization back later, but while refactoring it's
safer to always recompute the transformation.