texture: apply texture matrices after texture coordinate generation
The two operations are not mutually exclusive: first, go through the
texture coordinate generation (if disabled, then the identity matrix is
applied to the vertex texture coordinates), and then the texture matrix
is applied to the result.

The mechanism provided by the GX_SetTexCoordGen2() function seems to fix
well in this case, since it allows for two texture matrices to be
applied one after the other: we use the first one for texture
coordinate generation, the second for the texture transformation matrix.
Should this turn out to be a bad idea, we could pre-multiply the two
matrices in software.