TEV: store raster color output in a TEV register
The GX_CC_RASC and GX_CA_RASA constants do not refer to the untextured
vertex color, but only to the color being generated by the color channel
associated with the current stage: while in some situation these two
entities coincide, in the case of a lit object we do have two active
color channels, and these constants can only be used to refer to one of
them at a time. Furthermore, when we enter the texturing stage on a lit
geometry, we pass GX_COLORNULL, meaning that no rasterizer color channel
is associated, and here GX_CC_RASC and GX_CA_RASA only render black.

Yet, the texture environment operations in OpenGL sometimes need to
operate on the color emitted on the vertex by the rasterizer.

The solution is to emit the result of the lighting operations not on
GX_TEVPREV (which gets changed at each stage), but on a well-named
register (one of GX_TEVREG0 - GX_TEVREG2), which later stages can
then refer to.