lighting: always take alpha from diffuse color
This is a fixup of commit 470b0272f37d877c379f768f0e51becfdd97eafc: the
OpenGL spec says that the alpha value resulting from the raster
operations is always the alpha value of the diffuse colour. We compute
it in the second TEV stage, so modify the TEV operation to always use
GX_CA_RASA without mixing it with any other value.