points: support point texturing
If the GL_COORD_REPLACE flag is set, then instead of the texture
coordinates found on the vertex, we should use a hardcoded (0, 0)
coordinate for the top left corner, and (1, 1) for the bottom right.
We do this by replacing the vertex array with the PointSpritesTexReader
class, whose reader method always returns (0, 0).

We also need to modify a few other places so that the array classes know
what is the primitive being drawn, because this swap must only be
applied when rendering points.