state: setup fog only if its settings changed
Move the setup_fog() sunction out of the _ogx_setup_render_stages()
function and into _ogx_apply_state(), since by render stages we mean all
waht is related with the TEV stages.

Add a dirty bit to signal when the fogging configuration has changed and
needs to be setup again.