remove bogus nehe lesson
While this code may have been useful for initial testing of the library it is not a good example of how to construct wii/gamecube homebrew projects. Use of the template Makefiles or the more recent cmakery should be preferred and, ideally, reworking code like this to use embedded data should be avoided.