arrays: store VBO at array definition time
The association of a VBO to an array is to be saved in the client state:
we should look up the VBO only when we draw some geometry. The error was
discovered running the cube_tex example (it will be added to the repo in
a later commit), which uses two different VBOs for the position and
texture coordinate attributes.