stencil: mark the TEV as dirty after drawing to the stencil buffer
Fixes a regression introduced by commit 836fd7e7e8d4, because indeed
the _ogx_stencil_draw() function (and draw_op() in particular) alters
the TEV setup, which then needs to be restored when drawing to the
scene.

This fixes an issue in the dinoshade program, where some parts of the
reflected dinosaur could be seen outside of the floor.