shader: add a draw_done callback
Sometimes the shaders need to change some GX parameters that are not
touched by OpenGX itself and which need to be reverted after the drawing
commands have been sent, or risk altering the following drawing
operations. Swap tables and indirect rendering settings are an example
of such parameters.

Therefore, allow the client to restore a post drawing function, where
all these settings can be reverted to their default values.