arrays: support passing colors with just 1 or 2 components
This is unusual for OpenGL < 2.0, but when using shaders, it happens
that attributes have only one or two components. We can always use
texture coordinates in taht case (since these have one or two
components), but colors are easier to manipulate in the TEV, because
they don't require a texture.