VBOs: make sure that VBO data is not altered or free while in use
When issuing drawing commands whose vertex data are stored in VBOs, we
can not alter the contents of the VBOs until the drawing operation has
finished.

We implement this by sending a "draw sync" token to the GX pipeline, and
installing a draw sync callback function to keep track of the last token
which has been drawn. We use progressive numbers as tokens, so taht we
can use a simple arithmetical comparison to detect if the draw operation
we are interested in has been completed or not.

In case glBufferSubData() is called on a VBO which is currently in use,
we have to stall the pipeline and busy-loop until the token has been
received. This is not optimal, but it's also how the glBufferSubData()
command is documented to work in OpenGL.

For glBufferData(), instead, since we are going to make a new allocation
in any case, we can avoid stalling and instead preserve the old VBO
buffers until they have completed drawing. We do this by keeping them in
a list of unbound buffers, and periodically clean them up: either
checking if they have completed every time glBufferData() is called, or
unconditionally deleting them whenever the frame is complete.

This fixes an issue in the Neverball game, where coins where
occasionally disappearing.